requirements.txt
pyobjc>=9.0
Pillow>=9.0
pyperclip>=1.8.2
flask>=3.0